+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Updated `piral-cli-webpack5` to never output empty `main.css` files in pilets
 - Updated `piral-oidc` to use up-to-date `oidc-client-ts` instead of discontinued `oidc-client` library (#769)
 - Added support for Angular 20 in `piral-ng`
+- Added warning in `piral-cli` when shared dependencies are wrongly declared (#768)
 - Added support for Promise-based Webpack configuration files in `piral-cli-webpack5` (#774) by @grant-progress
 - Added `remoteTypes` to *piral.json* for specifying extra declarations